# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rewards API",
    "description": "API-first rewards service built with TypeScript, Fastify and SOLID principles",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/persons/{id}/points": {
      "get": {
        "tags": [
          "points"
        ],
        "description": "List the most recent points ledger entries for a person. Providers may view anyone, receivers only themselves",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "offset",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "personId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "entries": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "personId": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "earned",
                              "removed",
                              "redeemed"
                            ]
                          },
                          "points": {
                            "type": "integer",
                            "exclusiveMinimum": true,
                            "maximum": 9007199254740991,
                            "minimum": 0
                          },
                          "reason": {
                            "type": "string"
                          },
                          "balanceAfter": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "personId",
                          "type",
                          "points",
                          "balanceAfter",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "total": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "limit": {
                      "type": "integer",
                      "exclusiveMinimum": true,
                      "maximum": 9007199254740991,
                      "minimum": 0
                    },
                    "offset": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "hasMore": {
                      "type": "boolean"
                    },
                    "appVersion": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "personId",
                    "entries",
                    "total",
                    "limit",
                    "offset",
                    "hasMore",
                    "appVersion",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "appVersion": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "appVersion",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "appVersion": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "appVersion",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/points/add": {
      "post": {
        "tags": [
          "points"
        ],
        "description": "Grant loyalty points to a person. The authenticated provider is the grantor, and a provider cannot grant points to themselves",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "personId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "points": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "maximum": 9007199254740991,
                    "minimum": 0
                  },
                  "reason": {
                    "type": "string",
                    "maxLength": 200
                  }
                },
                "required": [
                  "personId",
                  "points"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "personId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "earned",
                        "removed",
                        "redeemed"
                      ]
                    },
                    "points": {
                      "type": "integer",
                      "exclusiveMinimum": true,
                      "maximum": 9007199254740991,
                      "minimum": 0
                    },
                    "balance": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "appVersion": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "personId",
                    "type",
                    "points",
                    "balance",
                    "appVersion",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "appVersion": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "appVersion",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "appVersion": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "appVersion",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "appVersion": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "appVersion",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/points/remove": {
      "post": {
        "tags": [
          "points"
        ],
        "description": "Remove loyalty points from a person. Only a provider can remove points",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "personId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "points": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "maximum": 9007199254740991,
                    "minimum": 0
                  },
                  "reason": {
                    "type": "string",
                    "maxLength": 200
                  }
                },
                "required": [
                  "personId",
                  "points"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "personId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "earned",
                        "removed",
                        "redeemed"
                      ]
                    },
                    "points": {
                      "type": "integer",
                      "exclusiveMinimum": true,
                      "maximum": 9007199254740991,
                      "minimum": 0
                    },
                    "balance": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "appVersion": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "personId",
                    "type",
                    "points",
                    "balance",
                    "appVersion",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "appVersion": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "appVersion",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "appVersion": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "appVersion",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "appVersion": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "appVersion",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/points/redeem": {
      "post": {
        "tags": [
          "points"
        ],
        "description": "Redeem a person’s loyalty points",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "personId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "points": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "maximum": 9007199254740991,
                    "minimum": 0
                  },
                  "reason": {
                    "type": "string",
                    "maxLength": 200
                  }
                },
                "required": [
                  "personId",
                  "points"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "personId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "earned",
                        "removed",
                        "redeemed"
                      ]
                    },
                    "points": {
                      "type": "integer",
                      "exclusiveMinimum": true,
                      "maximum": 9007199254740991,
                      "minimum": 0
                    },
                    "balance": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "appVersion": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "personId",
                    "type",
                    "points",
                    "balance",
                    "appVersion",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "appVersion": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "appVersion",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "appVersion": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "appVersion",
                    "timestamp"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "servers": [
    {
      "url": "http://localhost:13002",
      "description": "Rewards API"
    }
  ]
}